Saeeam django_project/requirements.txt
asgiref==3.2.10
Django==3.1.1
django-ckeditor==6.0.0
django-crispy-forms==1.9.2
django-js-asset==1.2.2
django-phonenumber-field==5.0.0
phonenumbers==8.12.10
Pillow==7.2.0
PyMySQL==0.10.1
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
